# api_gateway/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GDI MCP Server API",
    "version": "1.0.0",
    "description": "API de Gestión Documental Inteligente para gobiernos latinoamericanos. Permite consultar expedientes, documentos, firmas pendientes y catálogos del sistema."
  },
  "servers": [
    {
      "url": "__OPENAPI_BASE_URL__"
    }
  ],
  "paths": {
...
    "/api/v1/documents/{document_id}/reject": {
      "post": {
        "operationId": "rejectDocument",
        "summary": "Rechazar documento",
        "description": "Rechaza un documento en proceso de firma",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "UUID del documento"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Motivo del rechazo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Documento rechazado"
          },
          "403": {
            "description": "No autorizado"
          },
          "409": {
            "description": "Estado invalido"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/documents/{document_id}/signature-details": {
      "get": {
        "operationId": "getSignatureDetails",
        "summary": "Detalles de firma de un documento",
        "description": "Obtiene lista de firmantes, estado de cada uno y orden de firma",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "UUID del documento"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalles de firma con firmantes y estados"
          },
          "404": {
            "description": "Documento no encontrado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/records/{record_id}/fields/{field_name}": {
      "patch": {
        "operationId": "updateRecordField",
        "summary": "Actualizar un campo individual del legajo",
        "description": "Actualiza un solo campo dentro del data JSONB del legajo.",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campo actualizado"
          },
          "404": {
            "description": "Legajo o campo no encontrado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/records/{record_id}/fields/{field_name}/verify": {
      "post": {
        "operationId": "verifyRecordField",
        "summary": "Marcar campo como verificado",
        "description": "Marca un campo del legajo como verificado con timestamp y usuario.",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "field_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Campo verificado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/v1/records/{record_id}/history": {
      "get": {
        "operationId": "getRecordHistory",
        "summary": "Historial del legajo",
        "description": "Devuelve el historial de cambios del legajo enriquecido con contexto de documentos y expedientes vinculados.",
        "parameters": [
          {
            "name": "record_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Historial del legajo"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/memos/received": {
      "get": {
        "operationId": "getReceivedMemos",
        "summary": "Memos recibidos",
        "description": "Obtiene memos recibidos por el usuario autenticado",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Página"
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Resultados por página (max 100)"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Texto a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de memos recibidos"
          },
          "401": {
            "description": "No autorizado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/memos/archived": {
      "get": {
        "operationId": "getArchivedMemos",
        "summary": "Memos archivados",
        "description": "Obtiene memos archivados por el usuario autenticado",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Página"
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Resultados por página (max 100)"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Texto a buscar"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de memos archivados"
          },
          "401": {
            "description": "No autorizado"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/v1/search/semantic": {
      "get": {
        "operationId": "semanticSearch",
        "summary": "Búsqueda semántica de documentos",
        "description": "Busca documentos similares por contenido usando embeddings, con vinculaciones a expedientes y legajos",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Texto de búsqueda (mínimo 3 caracteres)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Máximo de resultados (max 50)"
          }
        ],
        "responses": {
          "200": {
            "description": "Documentos similares con vinculaciones a expedientes y legajos"
          },
          "400": {
            "description": "query requerido (min 3 caracteres)"
          },
          "503": {
            "description": "Búsqueda temporalmente no disponible"
          }
        },
        "security": [
          {
            "OAuth2": [
              "openid",
              "profile",
              "email"
            ]
          },
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OAuth2": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://__OPENAPI_AUTH0_DOMAIN__/authorize",
            "tokenUrl": "https://__OPENAPI_AUTH0_DOMAIN__/oauth/token",
            "scopes": {
              "openid": "OpenID Connect",
              "profile": "User profile",
              "email": "User email"
            }
          }
        }
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key para autenticación (fallback)"
      }
    }
  }
}